requirements.txt
# API
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.2
python-dotenv==1.0.0
starlette==0.27.0

# Security
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
ratelimit==2.2.1

# Image Processing
Pillow==10.1.0
numpy==1.26.2
opencv-python==4.8.1.78
torch==2.1.1
torchvision==0.16.1
scikit-image==0.22.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
aioredis==2.0.1

# Storage
boto3==1.29.6
azure-storage-blob==12.19.0

# Monitoring
prometheus-client==0.19.0
sentry-sdk==1.35.0
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
mypy==1.7.1
flake8==6.1.0
pre-commit==3.5.0
